+++ tests/test_generator.py
@@ -12,12 +12,12 @@ class TestGenerateCore(unittest.TestCase):
         g = Scru128Generator()
         self.assertEqual(g._last_status, Scru128Generator.Status.NOT_EXECUTED)
 
-        prev = g.generate_core(ts)
+        prev = g.generate_core(ts, 10_000)
         self.assertEqual(g._last_status, Scru128Generator.Status.NEW_TIMESTAMP)
         self.assertEqual(prev.timestamp, ts)
 
         for i in range(100_000):
-            curr = g.generate_core(ts - min(9_998, i))
+            curr = g.generate_core(ts - min(9_998, i), 10_000)
             self.assertTrue(
                 g._last_status == Scru128Generator.Status.COUNTER_LO_INC
                 or g._last_status == Scru128Generator.Status.COUNTER_HI_INC
@@ -33,17 +33,17 @@ class TestGenerateCore(unittest.TestCase):
         g = Scru128Generator()
         self.assertEqual(g._last_status, Scru128Generator.Status.NOT_EXECUTED)
 
-        prev = g.generate_core(ts)
+        prev = g.generate_core(ts, 10_000)
         self.assertEqual(g._last_status, Scru128Generator.Status.NEW_TIMESTAMP)
         self.assertEqual(prev.timestamp, ts)
 
-        curr = g.generate_core(ts - 10_000)
+        curr = g.generate_core(ts - 10_000, 10_000)
         self.assertEqual(g._last_status, Scru128Generator.Status.CLOCK_ROLLBACK)
         self.assertGreater(prev, curr)
         self.assertEqual(curr.timestamp, ts - 10_000)
 
         prev = curr
-        curr = g.generate_core(ts - 10_001)
+        curr = g.generate_core(ts - 10_001, 10_000)
         self.assertTrue(
             g._last_status == Scru128Generator.Status.COUNTER_LO_INC
             or g._last_status == Scru128Generator.Status.COUNTER_HI_INC
@@ -59,13 +59,13 @@ class TestGenerateCoreNoRewind(unittest.TestCase):
         g = Scru128Generator()
         self.assertEqual(g._last_status, Scru128Generator.Status.NOT_EXECUTED)
 
-        prev = g.generate_core_no_rewind(ts)
+        prev = g.generate_core_no_rewind(ts, 10_000)
         assert prev is not None
         self.assertEqual(g._last_status, Scru128Generator.Status.NEW_TIMESTAMP)
         self.assertEqual(prev.timestamp, ts)
 
         for i in range(100_000):
-            curr = g.generate_core_no_rewind(ts - min(9_998, i))
+            curr = g.generate_core_no_rewind(ts - min(9_998, i), 10_000)
             assert curr is not None
             self.assertTrue(
                 g._last_status == Scru128Generator.Status.COUNTER_LO_INC
@@ -82,16 +82,16 @@ class TestGenerateCoreNoRewind(unittest.TestCase):
         g = Scru128Generator()
         self.assertEqual(g._last_status, Scru128Generator.Status.NOT_EXECUTED)
 
-        prev = g.generate_core_no_rewind(ts)
+        prev = g.generate_core_no_rewind(ts, 10_000)
         assert prev is not None
         self.assertEqual(g._last_status, Scru128Generator.Status.NEW_TIMESTAMP)
         self.assertEqual(prev.timestamp, ts)
 
-        curr = g.generate_core_no_rewind(ts - 10_000)
+        curr = g.generate_core_no_rewind(ts - 10_000, 10_000)
         assert curr is None
         self.assertEqual(g._last_status, Scru128Generator.Status.NEW_TIMESTAMP)
 
-        curr = g.generate_core_no_rewind(ts - 10_001)
+        curr = g.generate_core_no_rewind(ts - 10_001, 10_000)
         assert curr is None
         self.assertEqual(g._last_status, Scru128Generator.Status.NEW_TIMESTAMP)